Snapshot of logical keyboard experimentation
This is chock full of hacks but the stray capitalisation produced by
typing with shift held down seems to be at least mostly fixed (I'm not
saying it's perfect). The most noticeable flaw now is that there is a
tendency to lose some characters. I want to commit this before fiddling
further.